Allow correct user to delete posts.

+++ app/routes.py
@@ -172,23 +172,33 @@ def edit_post(id):
     return render_template('edit_post.html',form=form)
 
 @app.route("/posts/delete/<int:id>")
+@login_required
 def delete_post(id):
     post_to_delete = Post.query.get_or_404(id)
-    try:
-        db.session.delete(post_to_delete)
-        db.session.commit()
+    id = current_user.id
+    if id == post_to_delete.poster_id:
+        try:
+            db.session.delete(post_to_delete)
+            db.session.commit()
 
+            # Return Message
+            flash("Post was successfully deleted!")
+            
+            # Get All Posts from DB
+            posts = Post.query.order_by(Post.date_posted)
+            return render_template("posts.html", posts=posts)
+        except:
+            flash("Error deleting post, try again")
+            # Get All Posts from DB
+            posts = Post.query.order_by(Post.date_posted)
+            return render_template("posts.html", posts=posts) 
+    else:
         # Return Message
-        flash("Post was successfully deleted!")
+        flash("You are not authorized to delete this post.")
         
         # Get All Posts from DB
         posts = Post.query.order_by(Post.date_posted)
         return render_template("posts.html", posts=posts)
-    except:
-        flash("Error deleting post, try again")
-        # Get All Posts from DB
-        posts = Post.query.order_by(Post.date_posted)
-        return render_template("posts.html", posts=posts) 
 
 # Login
 @app.route("/login", methods=["GET", "POST"])
